+++ class-components/src/components/AppRouter.tsx
@@ -1,6 +1,7 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+
 import { HomePage, Page404 } from '../pages';
-import ListItemDetailed from './ListItemDetailed';
+import { ListItemDetailed } from './ListItemDetailed';
 
 const router = createBrowserRouter([
   {
@@ -18,4 +19,4 @@ const router = createBrowserRouter([
 
 const AppRouter = () => <RouterProvider router={router} />;
 
-export default AppRouter;
+export { AppRouter };
