+++ class-components/src/components/Header.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 
 const Header = () => (
-  <header>
+  <header role="heading">
     <div>
       <Link to={'/'}>
         <img src="images/star-wars-logo.png" alt="Star Wars logo" />
@@ -11,4 +11,4 @@ const Header = () => (
   </header>
 );
 
-export default Header;
+export { Header };
