+++ class-components/src/components/Loader.tsx
@@ -4,4 +4,4 @@ const Loader = () => (
   </div>
 );
 
-export default Loader;
+export { Loader };
