+++ class-components/src/components/__tests__/ErrorBoundary.test.tsx
@@ -1,6 +1,6 @@
 import { screen, render, fireEvent } from '@testing-library/react';
 
-import ErrorBoundary from '../ErrorBoundary';
+import { ErrorBoundary } from '../ErrorBoundary';
 
 describe('ErrorBoundary', () => {
   it('should render errorBoundary when throwing an error', () => {
