+++ class-components/src/components/__tests__/Header.test.tsx
@@ -1,8 +1,7 @@
 import { screen } from '@testing-library/react';
 
 import { renderWithRouter } from '../../utils';
-
-import Header from '../Header';
+import { Header } from '../Header';
 
 describe('Header', () => {
   it('should render correctly', () => {
