+++ class-components/src/components/__tests__/ListItem.test.tsx
@@ -1,8 +1,7 @@
 import { screen } from '@testing-library/react';
 
 import { renderWithRouter } from '../../utils';
-
-import ListItem from '../ListItem';
+import { ListItem } from '../ListItem';
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
