+++ class-components/src/components/__tests__/ListItemDetailed.test.tsx
@@ -1,8 +1,7 @@
 import { screen, waitForElementToBeRemoved } from '@testing-library/react';
 
 import { renderWithRouter } from '../../utils';
-
-import ListItemDetailed from '../ListItemDetailed';
+import { ListItemDetailed } from '../ListItemDetailed';
 import * as apiMock from '../../api';
 
 const details = {
