+++ class-components/src/components/__tests__/Pagination.test.tsx
@@ -1,8 +1,7 @@
 import { screen, fireEvent } from '@testing-library/react';
 
 import { renderWithRouter } from '../../utils';
-
-import Pagination from '../Pagination';
+import { Pagination } from '../Pagination';
 
 describe('Pagination', () => {
   it('should renders correctly with two buttons and set proper classNames', () => {
